Boolean calculus.

+++ common.c
@@ -907,7 +907,7 @@ boolean ispref)
 {
 name_pointer p= name_ptr;
 char*s= first_chunk(p);
-int name_len= (int)(last-first)+ispref;
+int name_len= (int)(last-first)+(int)ispref;
 if(s+name_len> byte_mem_end)overflow("byte memory");
 if(name_ptr+1>=name_dir_end)overflow("name");
 (++name_ptr)->byte_start= byte_ptr= s+name_len;
@@ -936,7 +936,7 @@ boolean ispref)
 {
 char*s;
 name_pointer q= p+1;
-int name_len= (int)(last-first)+ispref;
+int name_len= (int)(last-first)+(int)ispref;
 if(name_ptr>=name_dir_end)overflow("name");
 while(q->link!=name_dir)q= q->link;
 q->link= name_ptr;
